Refactor on tests from fract4d, fract4d_compiler and fract4dgui packages. (#56)
