+++ fract4d/tests/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+
+# package file for fract4d tests
+
