+++ fract4d/tests/test_colorizer.py
