+++ fract4d/tests/test_fract4d.py
@@ -8,7 +8,7 @@ from . import testbase
 
 from fract4d import fract4dc, gradient, image, messages
 
-from .test_fractalsite import FractalSite
+from .fractalsite import FractalSite
 
 pos_params = [
     0.0, 0.0, 0.0, 0.0,
@@ -25,11 +25,11 @@ class Test(testbase.ClassSetup):
         cf1 = self.compiler.get_formula("gf4d.cfrm","default","cf0")
         self.assertEqual(len(cf1.errors),0)
         self.compiler.compile(cf1)
-        
+
         cf2 = self.compiler.get_formula("gf4d.cfrm","zero","cf1")
         self.assertEqual(len(cf2.errors),0)
         self.compiler.compile(cf2)
-        
+
         f = self.compiler.get_formula("gf4d.frm","Mandelbrot")
 
         self.color_mandel_params = f.symbols.default_params() + \
@@ -45,7 +45,7 @@ class Test(testbase.ClassSetup):
         cf2 = self.compiler.get_formula("gf4d.cfrm","zero","cf1")
         self.assertEqual(len(cf2.errors),0)
         self.compiler.compile(cf2)
-        
+
         f = self.compiler.get_formula("test.frm","test_simpleshape")
         outputfile = self.compiler.compile_all(f,cf1,cf2,[])
 
@@ -54,7 +54,7 @@ class Test(testbase.ClassSetup):
                                      cf2.symbols.default_params()
 
         return outputfile
-    
+
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
@@ -65,7 +65,7 @@ class Test(testbase.ClassSetup):
         self.compiler.load_formula_file("gf4d.frm")
         self.compiler.load_formula_file("gf4d.cfrm")
         self.gradient = gradient.Gradient()
-        
+
     def tearDown(self):
         pass
 
@@ -86,23 +86,23 @@ class Test(testbase.ClassSetup):
         pfunc = fract4dc.pf_create(handle)
 
         fract4dc.pf_init(pfunc,pos_params, [self.gradient, 4.0, 0.5])
-        
+
         # a point which doesn't bail out
         result = fract4dc.pf_calc(pfunc,[0.15, 0.0, 0.0, 0.0],100,0,0,0)
         self.assertEqual(result,(100, 32, 0.0,0))
-        
+
         # one which does
         result = fract4dc.pf_calc(pfunc,[1.0, 1.0, 0.0, 0.0],100,0,0,0)
-        self.assertEqual(result,(1,0, 0.0,0)) 
+        self.assertEqual(result,(1,0, 0.0,0))
 
         # one which is already out
         result = fract4dc.pf_calc(pfunc,[17.5, 14.0, 0.0, 0.0],100,0,0,0)
-        self.assertEqual(result,(0, 0, 0.0,0)) 
+        self.assertEqual(result,(0, 0, 0.0,0))
 
         # without optional args
         result = fract4dc.pf_calc(pfunc,[17.5, 14.0, 0.0, 0.0],100)
-        self.assertEqual(result,(0, 0, 0.0,0)) 
-        
+        self.assertEqual(result,(0, 0, 0.0,0))
+
         pfunc = None
         handle = None
 
@@ -148,11 +148,11 @@ class Test(testbase.ClassSetup):
 
         (w,h,tw,th) = (40,20,40,20)
         im = image.T(w,h)
-        
+
         cmap = fract4dc.cmap_create([(1.0, 255, 255, 255, 255)])
 
         fw = fract4dc.fw_create(1,pfunc,cmap,im._img,site)
-        
+
         ff = fract4dc.ff_create(
             [0.0, 0.0, 0.0, 0.0,
              4.0,
@@ -185,9 +185,9 @@ class Test(testbase.ClassSetup):
         # check they are updated if image is bigger
         (w,h,tw,th) = (40,20,400,200)
         im = image.T(w,h,tw,th)
-        
+
         fw = fract4dc.fw_create(1,pfunc,cmap,im._img,site)
-        
+
         ff = fract4dc.ff_create(
             [0.0, 0.0, 0.0, 0.0,
              4.0,
@@ -222,7 +222,7 @@ class Test(testbase.ClassSetup):
         im.set_offset(offx, offy)
 
         fw = fract4dc.fw_create(1,pfunc,cmap,im._img,site)
-        
+
         ff = fract4dc.ff_create(
             [0.0, 0.0, 0.0, 0.0,
              4.0,
@@ -253,31 +253,31 @@ class Test(testbase.ClassSetup):
         self.assertNearlyEqual(topleft, [
             -2.0 + dx[0] * (offx + 0.5),
              1.0 + dy[1] * (offy + 0.5),
-            0.0,0.0])        
-                
+            0.0,0.0])
+
     def testFractWorker(self):
         xsize = 8
         ysize = 8
         im = image.T(xsize,ysize)
-        
+
         cmap = fract4dc.cmap_create([(1.0, 255, 255, 255, 255)])
 
         fract4dc.cmap_set_solid(cmap,0,0,0,0,255)
         fract4dc.cmap_set_solid(cmap,1,0,0,0,255)
-        
+
         (fw,ff,site,handle,pfunc) = self.makeWorkerAndFunc(im._img,cmap)
 
         im.clear()
         fate_buf = im.fate_buffer()
-        buf = im.image_buffer() 
-        
+        buf = im.image_buffer()
+
         # draw 1 pixel, check it's set properly
         fract4dc.fw_pixel(fw,0,0,1,1)
         self.assertPixelIs(im,0,0,[im.OUT]+[im.UNKNOWN]*3)
 
         fract4dc.fw_pixel(fw,0,4,1,1)
         self.assertPixelIs(im,0,4,[im.IN]+[im.UNKNOWN]*3)
-        
+
         # draw it again, check no change.
         fract4dc.fw_pixel(fw,0,0,1,1)
         self.assertPixelIs(im,0,0,[im.OUT]+[im.UNKNOWN]*3)
@@ -291,7 +291,7 @@ class Test(testbase.ClassSetup):
         cmap = fract4dc.cmap_create(
             [(1.0, 79, 88, 41, 255)])
         fract4dc.cmap_set_solid(cmap,1,100,101,102,255)
-        
+
         (fw,ff,site,handle,pfunc) = self.makeWorkerAndFunc(im._img,cmap)
 
         fract4dc.fw_pixel(fw,0,0,1,1)
@@ -311,7 +311,7 @@ class Test(testbase.ClassSetup):
 
         self.assertPixelIs(
             im,3,1, [im.UNKNOWN]*4,
-            [79,88,41], [100,101,102], im.OUT)        
+            [79,88,41], [100,101,102], im.OUT)
 
     def testMultiThreadedCalc(self):
         xsize = 64
@@ -345,7 +345,7 @@ class Test(testbase.ClassSetup):
             site=site)
 
         self.assertEqual(siteobj.progress_list[-1], 0.0)
-        self.assertEqual(siteobj.progress_list[-2], 1.0)                        
+        self.assertEqual(siteobj.progress_list[-2], 1.0)
 
         self.assertTrue(siteobj.image_list[-1]==(0,0,xsize,ysize))
 
@@ -365,7 +365,7 @@ class Test(testbase.ClassSetup):
                     (i % (im.FATE_SIZE * xsize)) // im.FATE_SIZE,
                     i // (im.FATE_SIZE * xsize),
                     i % im.FATE_SIZE)
-            
+
             if i % 4 == 0:
                 # no-aa
                 self.assertNotEqual(byte, 255,
@@ -378,9 +378,9 @@ class Test(testbase.ClassSetup):
         self.assertPixelCount(xsize,ysize,siteobj)
 
     def assertPixelCount(self,xsize,ysize,siteobj):
-        # total pixels calculated should == w*h        
+        # total pixels calculated should == w*h
         self.assertEqual(xsize*ysize,siteobj.stats_list[-1].pixels)
-        for stats in siteobj.stats_list:            
+        for stats in siteobj.stats_list:
             # pixels calced + pixels skipped = pixels
             self.assertEqual(
                 stats.pixels,stats.pixels_calculated + stats.pixels_skipped)
@@ -420,7 +420,7 @@ class Test(testbase.ClassSetup):
             site=site)
 
         self.assertEqual(siteobj.progress_list[-1], 0.0)
-        self.assertEqual(siteobj.progress_list[-2], 1.0)                        
+        self.assertEqual(siteobj.progress_list[-2], 1.0)
 
         self.assertTrue(siteobj.image_list[-1]==(0,0,xsize,ysize))
 
@@ -440,7 +440,7 @@ class Test(testbase.ClassSetup):
                     (i % (im.FATE_SIZE * xsize)) // im.FATE_SIZE,
                     i // (im.FATE_SIZE * xsize),
                     i % im.FATE_SIZE)
-            
+
             if i % 4 == 0:
                 # no-aa
                 self.assertNotEqual(byte, 255,
@@ -515,7 +515,7 @@ class Test(testbase.ClassSetup):
 
         vec = fract4dc.eye_vector(params,1.0)
         self.assertEqual(vec, (-0.0, -0.0, -1.0, -0.0))
-        
+
         params[6] = math.pi/2.0
         mat = fract4dc.rot_matrix(params)
         self.assertNearlyEqual(mat, ((0.0, 0.0, 1.0, 0.0),
@@ -525,7 +525,7 @@ class Test(testbase.ClassSetup):
 
         vec = fract4dc.eye_vector(params,10.0)
         self.assertNearlyEqual(vec, (10.0, -0.0, -0.0, -0.0))
-                        
+
     def testFDSite(self):
         xsize = 64
         ysize = int(xsize * 3.0/4.0)
@@ -566,7 +566,7 @@ class Test(testbase.ClassSetup):
                 if nrecved == x:
                     #print "hit message count"
                     fract4dc.interrupt(site)
-                
+
                 nb = 2*4
                 bytes = os.read(rfd,nb)
                 if len(bytes) < nb:
@@ -581,16 +581,16 @@ class Test(testbase.ClassSetup):
                 if len(bytes) < size:
                     self.fail("bad message")
                     break
-                
+
                 msg = messages.parse(t, bytes)
                 #print "msg: %s" % msg.show()
                 if msg.name == "Status" and msg.status == 0:
                     #done
                     #print "done"
                     break
-                
+
                 nrecved += 1
-            
+
     def testDirtyFlagFullRender(self):
         '''Render the same image 2x with different colormaps.
 
@@ -623,7 +623,7 @@ class Test(testbase.ClassSetup):
                 self.assertEqual(a,b)
             i += 1
 
-        
+
     def drawTwice(self,is_dirty,xsize):
         ysize = int(xsize * 3.0/4.0)
         im = image.T(xsize,ysize)
@@ -637,7 +637,7 @@ class Test(testbase.ClassSetup):
 
         cmap = fract4dc.cmap_create(
             [(1.0, 255, 255, 255, 255)])
-        
+
         fract4dc.calc(
             params=[0.0, 0.0, 0.0, 0.0,
              4.0,
@@ -658,7 +658,7 @@ class Test(testbase.ClassSetup):
         #print "1st pass %s" % is_dirty
         #fract4dc.image_save(image, "/tmp/pass1%d.tga" % is_dirty)
         #self.print_fates(image,xsize,ysize)
-        
+
         cmap = fract4dc.cmap_create(
             [(1.0, 76, 49, 189, 255)])
 
@@ -682,9 +682,9 @@ class Test(testbase.ClassSetup):
         #print "2nd pass %s" % is_dirty
         #self.print_fates(image,xsize,ysize)
         im.save(os.path.join(Test.tmpdir.name, "pass2%d.tga" % is_dirty))
-        
+
         return [] # fract4dc.image_buffer(image)
-        
+
     def testMiniTextRender(self):
         self.compileMandel()
         handle = fract4dc.pf_load(Test.pf_name)
@@ -703,8 +703,8 @@ class Test(testbase.ClassSetup):
         printable_image = "\n".join(image)
         self.assertEqual(printable_image[0], " ", printable_image)
         self.assertEqual(printable_image[20*41+20],"#", printable_image) # in the middle
-        #print printable_image # shows low-res mbrot in text mode 
-        
+        #print printable_image # shows low-res mbrot in text mode
+
     def testBadLoad(self):
         # wrong arg type/number
         self.assertRaises(TypeError,fract4dc.pf_load,1)
@@ -733,7 +733,7 @@ class Test(testbase.ClassSetup):
         handle = fract4dc.pf_load(Test.pf_name)
         pfunc = fract4dc.pf_create(handle)
         fract4dc.pf_init(pfunc, pos_params, [1,2,3,4])
-        
+
     def testBadCalc(self):
         self.compileMandel()
         handle = fract4dc.pf_load(Test.pf_name)
@@ -761,14 +761,14 @@ class Test(testbase.ClassSetup):
         self.assertEqual(fract4dc.cmap_lookup(cmap,1.0), (0,255,50,255))
         self.assertEqual(fract4dc.cmap_lookup(cmap,0.5), (127,127,75,255))
         self.assertEqual(fract4dc.cmap_lookup(cmap,0.00000001), (254,0,99,255))
-        
+
         cmap = fract4dc.cmap_create(
             [(0.0,255,0,100,255)])
         expc1 = (255,0,100,255)
         self.assertEqual(fract4dc.cmap_lookup(cmap,0.0),expc1)
         self.assertEqual(fract4dc.cmap_lookup(cmap,1.0),expc1)
         self.assertEqual(fract4dc.cmap_lookup(cmap,0.4),expc1)
-        
+
         colors = []
         for i in range(256):
             colors.append((i/255.0,(i*17)%256,255-i,i//2,i//2+127))
@@ -776,7 +776,7 @@ class Test(testbase.ClassSetup):
         cmap = fract4dc.cmap_create(colors)
         for i in range(256):
             self.assertEqual(fract4dc.cmap_lookup(cmap,i/255.0),colors[i][1:],i)
-            
+
     def testTransfers(self):
         # test fates
         cmap = fract4dc.cmap_create(
@@ -784,38 +784,38 @@ class Test(testbase.ClassSetup):
 
         # make inner transfer func none
         fract4dc.cmap_set_transfer(cmap,1,0)
-        
+
         # inside should be all-black by default, outside should never be
         index = 0.0
-        while index < 2.0: 
+        while index < 2.0:
             color = fract4dc.cmap_lookup_flags(cmap,index,0,1)
             self.assertEqual(color,(0,0,0,255))
             color = fract4dc.cmap_lookup_flags(cmap,index,0,0)
-            self.assertEqual(color,(33,33,33,255))            
+            self.assertEqual(color,(33,33,33,255))
             index += 0.1
 
         # test setting solid colors and transfers
         fract4dc.cmap_set_solid(cmap,0,166,166,166,255)
         fract4dc.cmap_set_solid(cmap,1,177,177,177,255)
         fract4dc.cmap_set_transfer(cmap,0,0)
-        
+
         index = 0.0
-        while index < 2.0: 
+        while index < 2.0:
             color = fract4dc.cmap_lookup_flags(cmap,index,0,1)
             self.assertEqual(color,(177,177,177,255))
             color = fract4dc.cmap_lookup_flags(cmap,index,0,0)
-            self.assertEqual(color,(166,166,166,255))            
+            self.assertEqual(color,(166,166,166,255))
             index += 0.1
 
         # make inner linear
         fract4dc.cmap_set_transfer(cmap,1,1)
 
         index = 0.0
-        while index < 2.0: 
+        while index < 2.0:
             color = fract4dc.cmap_lookup_flags(cmap,index,0,1)
             self.assertEqual(color,(33,33,33,255))
             color = fract4dc.cmap_lookup_flags(cmap,index,0,0)
-            self.assertEqual(color,(166,166,166,255))            
+            self.assertEqual(color,(166,166,166,255))
             index += 0.1
 
         # test that solid overrides
@@ -834,7 +834,7 @@ class Test(testbase.ClassSetup):
 
         (r2,g2,b2,a3) = fract4dc.hsl_to_rgb(eh,es,ev,a)
         self.assertEqual((r2,g2,b2,a3),(r,g,b,a))
-        
+
     def testColorTransformHSL(self):
         # black
         self.assertColorTransformHSL(
@@ -865,7 +865,7 @@ class Test(testbase.ClassSetup):
         self.assertColorTransformHSL(
             0.0,0.0,1.0,
             4.0,1.0,0.5)
-        
+
         # cyan
         self.assertColorTransformHSL(
             0.0,1.0,1.0,
@@ -881,7 +881,7 @@ class Test(testbase.ClassSetup):
             1.0,1.0,0.0,
             1.0,1.0,0.5)
 
-        
+
     def testColorTransformHSV(self):
         # red
         self.assertColorTransformHSV(
@@ -897,7 +897,7 @@ class Test(testbase.ClassSetup):
         self.assertColorTransformHSV(
             0.0,0.0,1.0,
             4.0,1.0,1.0)
-        
+
         # cyan
         self.assertColorTransformHSV(
             0.0,1.0,1.0,
@@ -932,21 +932,21 @@ class Test(testbase.ClassSetup):
         arena = fract4dc.arena_create(100,1)
         alloc = fract4dc.arena_alloc(arena, 1,1,10)
         alloc = fract4dc.arena_alloc(arena, 10,1,1)
-        
+
     def testTooSmallArena(self):
         self.assertRaises(MemoryError, fract4dc.arena_create,0,10)
         self.assertRaises(MemoryError, fract4dc.arena_create,10,0)
-        
+
     def testTooBigAlloc(self):
         arena = fract4dc.arena_create(10,1)
         self.assertRaises(MemoryError, fract4dc.arena_alloc,arena,8,1,10)
-        
+
     def testMultipleAllocs(self):
         arena = fract4dc.arena_create(10,1)
         for i in range(5):
             fract4dc.arena_alloc(arena,8,1,1)
 
-        # should be full now        
+        # should be full now
         self.assertRaises(MemoryError, fract4dc.arena_alloc,arena,8,1,1)
 
     def testReadArrayVal(self):
@@ -970,7 +970,7 @@ class Test(testbase.ClassSetup):
             val = i
             result = fract4dc.array_set_int(alloc,1, i,val)
             self.assertEqual(1,result)
-            
+
         for i in range(10):
             result = fract4dc.array_get_int(alloc,1, i)
             self.assertEqual(
@@ -1022,5 +1022,5 @@ class Test(testbase.ClassSetup):
                 self.assertEqual(
                     (i,1), val,
                     "bad result %s instead of %s for %d,%d" % (val,(i,1),x,y))
-                
+
                 i += 1
