+++ fract4d_compiler/tests/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+
+# subpackage file for fract4d_compiler package tests
+
