+++ fract4dgui/tests/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+
+# subpackage file for fract4dgui package tests
+
