+++ fract4dgui/tests/test_director.py
