+++ fract4dgui/tests/test_gtkfractal.py
