+++ fract4dgui/tests/test_main_window.py
