+++ fract4dgui/tests/test_model.py
