+++ fract4dgui/tests/test_painter.py
