+++ fract4dgui/tests/test_preferences.py
