+++ fract4dgui/tests/test_settings.py
