+++ setup.py
@@ -154,7 +154,7 @@ and includes a Fractint-compatible parser for your own fractal formulas.''',
     maintainer_email='edwin@bathysphere.org',
     keywords="edwin@bathysphere.org",
     url='http://github.com/edyoung/gnofract4d/',
-    packages=['fract4d_compiler', 'fract4d', 'fract4dgui'],
+    packages=['fract4d_compiler', 'fract4d', 'fract4dgui', 'fract4d_compiler/tests', 'fract4d/tests', 'fract4dgui/tests'],
     package_data={'fract4dgui': ['shortcuts-gnofract4d.ui', 'ui.xml']},
     ext_modules=modules,
     scripts=['gnofract4d'],
